Moved AutoSettableProperties functionality into a Trait
Created new Trait AutoSettablePropertiesTrait and removed equivalent functionality from classes that were using the two methods now in the trait. Largely untested. Need to add significantly more unit tests to make these kinds of changes safer.
