+++ src/Model/FIT/Field.php
@@ -2,10 +2,11 @@
 
 namespace App\Model\FIT;
 
-use function Symfony\Component\String\u;
+use App\Utility\AutoSettablePropertiesTrait;
 
 class Field
 {
+  use AutoSettablePropertiesTrait;
 
   protected $name;
   protected $value; // TODO: I don't know what value represents in a definition field in the CSV file
@@ -47,36 +48,4 @@ class Field
     $this->units = $units;
   }
 
-
-  /**
-   * Attempt to set object properties from an associative array
-   *
-   * @param array $a  An associatve array with keys matching the object vars for `$this`
-   */
-  public function setPropertiesFromArray($a)
-  {
-    $settable_properties = $this->getAutoSettableProperties();
-
-    foreach ($settable_properties as $property_name) {
-      if(array_key_exists($property_name, $a)){
-        // First look for $this->setProperty(), and then use "$this->property = ... "
-        $property_setter = 'set' . u($property_name)->camel()->title();
-        if(method_exists($this, $property_setter)){
-          $this->{$property_setter}($a[$property_name]);
-        } else {
-          $this->{$property_name} = $a[$property_name];
-        }
-      }
-    }
-  }
-
-  /**
-   * Returns the list of properties that can be set by setPropertiesFromArray
-   *
-   * Can be overwritten by children in order to provide more specific control
-   */
-  protected function getAutoSettableProperties()
-  {
-    return array_keys(get_object_vars($this));
-  }
 }
